// code-review-management/types/github.types.ts
export type Comment = z.infer<typeof CommentSchema>;
// ...
const side = ["LEFT", "RIGHT"] as const;
const subjectType = ["line", "file"] as const;
const authorAssociation = [
  "COLLABORATOR",
  "CONTRIBUTOR",
  "FIRST_TIMER",
  "FIRST_TIME_CONTRIBUTOR",
  "MANNEQUIN",
  "MEMBER",
  "NONE",
  "OWNER",
] as const;
const repoVisibility = ["public", "private", "internal"] as const;
const fileDiffStatus = [
  "added",
  "removed",
  "modified",
  "renamed",
  "copied",
  "changed",
  "unchanged",
] as const;
const compareCommitsStatus = [
  "diverged",
  "ahead",
  "behind",
  "identical",
] as const;

export const UserSchema = z.object({
  login: z.string(),
  id: z.number(),
  avatar_url: z.string(),
  permissions: z
    .object({
      pull: z.boolean(),
      push: z.boolean(),
      admin: z.boolean(),
      triage: z.boolean().optional(),
      maintain: z.boolean().optional(),
    })
    .optional(),
});
// ...
export const ReactionSchema = z.object({
  total_count: z.number(),
  "+1": z.number(),
  "-1": z.number(),
  laugh: z.number(),
  hooray: z.number(),
  confused: z.number(),
  heart: z.number(),
  rocket: z.number(),
  eyes: z.number(),
});

export const CommentSchema = z.object({
  pull_request_review_id: z.number(),
  id: z.number(),
  diff_hunk: z.string(),
  path: z.string(),
  commit_id: z.string(),
  original_commit_id: z.string(),
  user: UserSchema,
  body: z.string(),
  created_at: z.string(),
  updated_at: z.string(),
  reactions: ReactionSchema,
  start_line: z.number().nullable(),
  original_start_line: z.number().nullable(),
  start_side: z.enum(side).nullable(),
  line: z.number().nullable(),
  original_line: z.number().nullable(),
  side: z.enum(side),
  in_reply_to_id: z.number().nullish(),
  author_association: z.enum(authorAssociation),
  subject_type: z.enum(subjectType),
});